+++ app/src/main/java/io/punchtime/punchtime/ui/fragments/MonthFragment.java
@@ -1,10 +1,12 @@
-package io.punchtime.punchtime;
+package io.punchtime.punchtime.ui.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import io.punchtime.punchtime.R;
 
 /**
  * Created by haroenv on 26/03/16.
