+++ app/src/main/java/io/punchtime/punchtime/ui/fragments/ThreeDayFragment.java
@@ -1,4 +1,4 @@
-package io.punchtime.punchtime;
+package io.punchtime.punchtime.ui.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,14 +6,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import io.punchtime.punchtime.R;
+
 /**
  * Created by haroenv on 26/03/16.
  */
-public class WeekFragment extends Fragment {
+public class ThreeDayFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup parent, Bundle savedInstanceState) {
         // Defines the xml file for the fragment
-        return inflater.inflate(R.layout.fragment_week, parent, false);
+        return inflater.inflate(R.layout.fragment_three_day, parent, false);
     }
 
     // triggered soon after onCreateView
