+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="io.punchtime.punchtime.MainActivity"
+    tools:context=".ui.activities.MainActivity"
     tools:showIn="@layout/activity_main">
