+++ app/src/main/res/menu/main.xml
@@ -2,5 +2,5 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="io.punchtime.punchtime.MainActivity">
+    tools:context=".ui.activities.MainActivity">
 </menu>
